+++ docs/history.md
@@ -1,5 +1,10 @@
 # Racing+ Version History and News
 
+### *v0.78.0* - March 7th, 2022
+
+- In seeded races, players will now always get all Devil Rooms or all Angel Rooms. Each seed has a 50% chance of either option.
+- Fixed the bug where clean bedrooms would never appear in multi-character speedruns.
+
 ### *v0.77.2* - March 6th, 2022
 
 - The seeded death mechanic will now grant you one bone heart, one half red heart, and half a soul heart when you die on The Forgotten or The Soul. (Thanks Gamonymous)